requirements.txt
plotly
seaborn
numpy
streamlit
pandas
google-generativeai
python-dotenv
pathlib